+++ packages/cortex-react/docs/Input.stories.tsx
@@ -2,13 +2,10 @@ import { Meta, StoryObj } from '@storybook/react';
 import React, { useCallback, useState } from 'react';
 import { IoEye, IoPerson } from 'react-icons/io5';
 import {
-  CurrencyIMask,
-  ExpressionMasks,
-  Hint,
+  BRLMask,
   Input,
-  InputMaskEvent,
-  NumberIMask,
-  PercentageIMask,
+  Masks,
+  PercentageMask as PercentageMaskDef,
 } from '../src';
 
 export default {
@@ -87,7 +84,7 @@ export const Search: StoryObj<typeof Input.Search> = {
 export const ExpressionMask: StoryObj<typeof Input.Mask> = {
   args: {
     ...baseArgs,
-    mask: { mask: [ExpressionMasks.PHONE, ExpressionMasks.PHONE_EXTENDED] },
+    mask: Masks.COMBINED_PHONE,
     label: 'Expression Mask',
   },
   render: args => {
@@ -108,19 +105,18 @@ export const ExpressionMask: StoryObj<typeof Input.Mask> = {
 export const NumberMask: StoryObj<typeof Input.Mask> = {
   args: {
     ...baseArgs,
-    mask: NumberIMask,
     label: 'Number Mask',
   },
   render: args => {
-    const [numberValue, setNumberValue] = useState('1000');
+    const [numberValue, setNumberValue] = useState(1000);
 
     return (
       <Input.Mask
         placeholder={args.placeholder}
         label={args.label}
-        onChange={e => setNumberValue(e.unmaskedValue)}
+        onChange={e => setNumberValue(e)}
         value={numberValue}
-        mask={args.mask}
+        type="number"
       />
     );
   },
@@ -130,7 +126,7 @@ export const CurrencyMask: StoryObj<typeof Input.Mask> = {
   args: {
     ...baseArgs,
     label: 'Currency Mask',
-    mask: CurrencyIMask,
+    mask: BRLMask,
   },
   render: args => {
     const [currencyValue, setCurrencyValue] = useState('99,50');
@@ -139,7 +135,7 @@ export const CurrencyMask: StoryObj<typeof Input.Mask> = {
       <Input.Mask
         placeholder={args.placeholder}
         label={args.label}
-        onChange={e => setCurrencyValue(e.unmaskedValue)}
+        onChange={e => setCurrencyValue(e)}
         value={currencyValue}
         mask={args.mask}
       />
@@ -151,7 +147,7 @@ export const PercentageMask: StoryObj<typeof Input.Mask> = {
   args: {
     ...baseArgs,
     label: 'Percentage Mask',
-    mask: PercentageIMask,
+    mask: PercentageMaskDef,
   },
   render: args => {
     const [percentageValue, setPercentageValue] = useState('99,5');
@@ -160,7 +156,7 @@ export const PercentageMask: StoryObj<typeof Input.Mask> = {
       <Input.Mask
         placeholder={args.placeholder}
         label={args.label}
-        onChange={e => setPercentageValue(e.unmaskedValue)}
+        onChange={e => setPercentageValue(e)}
         value={percentageValue}
         mask={args.mask}
       />
@@ -168,77 +164,38 @@ export const PercentageMask: StoryObj<typeof Input.Mask> = {
   },
 };
 
-export const ShowUnmasked: StoryObj<typeof Input.Mask> = {
-  args: {
-    ...baseArgs,
-    mask: { mask: [ExpressionMasks.PHONE, ExpressionMasks.PHONE_EXTENDED] },
-    label: 'Expression Mask',
-  },
-  render: args => {
-    const [unmaskedValue, setUnmaskedValue] = useState('1112345678');
-    const [unmaskedNumberValue, setUnmaskedNumberValue] = useState('90,54');
-
-    return (
-      <div className="flex flex-col gap-y-micro">
-        <Input.Mask
-          placeholder={args.placeholder}
-          label={args.label}
-          onChange={e => {
-            setUnmaskedValue(e.unmaskedValue);
-          }}
-          value={unmaskedValue}
-          mask={args.mask}
-        />
-        <Hint>Unmasked: {unmaskedValue}</Hint>
-
-        <Input.Mask
-          label={'Number mask'}
-          onChange={e => {
-            setUnmaskedNumberValue(e.unmaskedValue);
-          }}
-          value={unmaskedNumberValue}
-          mask={CurrencyIMask}
-        />
-        <Hint>Unmasked: {unmaskedNumberValue}</Hint>
-      </div>
-    );
-  },
-};
-
-const total = 100;
-
 export const DoubleMask: StoryObj<typeof Input.Mask> = {
   args: {
     ...baseArgs,
     label: 'Expression Mask',
   },
   render: args => {
-    const [currency, setCurrency] = useState(String(total));
-    const [percentage, setPercentage] = useState(String(total));
+    const total = 540;
+    const [currency, setCurrency] = useState(0);
+    const [percentage, setPercentage] = useState(0);
 
     const handleChangeCurrency = useCallback(
-      (e: InputMaskEvent) => {
-        const rawNumber = Number(e.unmaskedValue ?? 0);
-        setCurrency(e.unmaskedValue);
+      (value: number) => {
+        const rawNumber = value ?? 0;
+        setCurrency(value);
         if (rawNumber === 0) {
-          setPercentage('0');
+          setPercentage(0);
         } else {
           const update = (rawNumber * 100) / total;
-          setPercentage(String(update));
+          setPercentage(update);
         }
       },
       [currency]
     );
 
     const handleChangePercentage = useCallback(
-      (e: InputMaskEvent) => {
-        const rawNumber = Number(e.unmaskedValue ?? 0);
-        setPercentage(e.unmaskedValue);
-        if (rawNumber === 0) {
-          setCurrency('0');
+      (value: number) => {
+        setPercentage(value);
+        if (value === 0) {
+          setCurrency(value);
         } else {
-          const update = (rawNumber / 100) * total;
-          setCurrency(String(update));
+          const update = (value / 100) * total;
+          setCurrency(update);
         }
       },
       [percentage]
@@ -246,19 +203,20 @@ export const DoubleMask: StoryObj<typeof Input.Mask> = {
 
     return (
       <>
+        <span>Total Amount: {total}</span>
         <Input.Mask
           placeholder={'currency'}
           label={'currency'}
           onChange={handleChangeCurrency}
           value={currency}
-          mask={CurrencyIMask}
+          mask={BRLMask}
         />
         <Input.Mask
           placeholder={'percentage'}
           label={'percentage'}
           onChange={handleChangePercentage}
           value={percentage}
-          mask={PercentageIMask}
+          mask={PercentageMaskDef}
         />
       </>
     );