+++ packages/cortex-react/src/tests/Input/Mask.test.tsx
@@ -1,17 +1,13 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
-import { CurrencyIMask, ExpressionMasks, Input, NumberIMask } from '../..';
+import { BRLMask, Input, Masks } from '../..';
 
 // IMPORTANT: fireEvent.blur after fireEvent.change needed to apply mask
 
 describe('InputMask', () => {
   describe('Expression', () => {
     it('Should apply correct expression mask on phone number masks', () => {
-      const maskOptions = {
-        mask: [ExpressionMasks.PHONE, ExpressionMasks.PHONE_EXTENDED],
-      };
-
-      render(<Input.Mask mask={maskOptions} />);
+      render(<Input.Mask mask={Masks.COMBINED_PHONE} />);
 
       const maskExpressionElement = screen.getByTestId(
         'input-box'
@@ -35,11 +31,7 @@ describe('InputMask', () => {
     });
 
     it('Should apply correct expression mask on cpf/cnpj masks', () => {
-      const maskOptions = {
-        mask: [ExpressionMasks.CPF, ExpressionMasks.CNPJ],
-      };
-
-      render(<Input.Mask mask={maskOptions} />);
+      render(<Input.Mask mask={Masks.COMBINED_CPF_CNPJ} />);
 
       const maskExpressionElement = screen.getByTestId(
         'input-box'
@@ -52,54 +44,12 @@ describe('InputMask', () => {
       fireEvent.blur(maskExpressionElement);
 
       expect(maskExpressionElement.value).toBe('123.456.789-10');
-
-      fireEvent.change(maskExpressionElement, {
-        target: { value: '12345678100001' },
-      });
-
-      fireEvent.blur(maskExpressionElement);
-
-      expect(maskExpressionElement.value).toBe('12.345.678/1000-01');
-    });
-  });
-
-  describe('Number', () => {
-    it('Should not render text on number mask', () => {
-      render(<Input.Mask mask={NumberIMask} />);
-
-      const maskNumberElement = screen.getByTestId(
-        'input-box'
-      ) as HTMLInputElement;
-
-      fireEvent.change(maskNumberElement, {
-        target: { value: 'teste... .' },
-      });
-
-      fireEvent.blur(maskNumberElement);
-
-      expect(maskNumberElement.value).toBe('');
-    });
-
-    it('Should apply number mask correctly', () => {
-      render(<Input.Mask mask={NumberIMask} />);
-
-      const maskNumberElement = screen.getByTestId(
-        'input-box'
-      ) as HTMLInputElement;
-
-      fireEvent.change(maskNumberElement, {
-        target: { value: '123456' },
-      });
-
-      fireEvent.blur(maskNumberElement);
-
-      expect(maskNumberElement.value).toBe('123456');
     });
   });
 
   describe('Currency', () => {
     it('Should apply currency mask', () => {
-      render(<Input.Mask mask={CurrencyIMask} />);
+      render(<Input.Mask mask={BRLMask} />);
 
       const maskCurrencyElement = screen.getByTestId(
         'input-box'
@@ -111,11 +61,11 @@ describe('InputMask', () => {
 
       fireEvent.blur(maskCurrencyElement);
 
-      expect(maskCurrencyElement.value).toBe('R$ 92,6');
+      expect(maskCurrencyElement.value).toBe('R$ 92,60');
     });
 
     it('Should not render text on currency mask', () => {
-      render(<Input.Mask mask={CurrencyIMask} />);
+      render(<Input.Mask mask={BRLMask} />);
 
       const maskCurrencyElement = screen.getByTestId(
         'input-box'
@@ -127,7 +77,7 @@ describe('InputMask', () => {
 
       fireEvent.blur(maskCurrencyElement);
 
-      expect(maskCurrencyElement.value).toBe('R$ ');
+      expect(maskCurrencyElement.value).toBe('R$ 0,00');
     });
   });
 });