+++ packages/react-charts/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.7-beta.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-charts@1.4.7-beta.0...@tecsinapse/react-charts@1.4.7-beta.1) (2024-08-21)
+
+**Note:** Version bump only for package @tecsinapse/react-charts
+
 ## [1.4.7-beta.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-charts@1.4.6...@tecsinapse/react-charts@1.4.7-beta.0) (2024-08-20)
 
 **Note:** Version bump only for package @tecsinapse/react-charts