+++ packages/react-native-kit/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.6-beta.1](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-native-kit@3.5.6-beta.0...@tecsinapse/react-native-kit@3.5.6-beta.1) (2024-08-21)
+
+**Note:** Version bump only for package @tecsinapse/react-native-kit
+
 ## [3.5.6-beta.0](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-native-kit@3.5.5...@tecsinapse/react-native-kit@3.5.6-beta.0) (2024-08-20)
 
 **Note:** Version bump only for package @tecsinapse/react-native-kit